requirements.txt
awpy==2.0.2
anthropic>=0.25.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pandas>=2.2.0
matplotlib>=3.9.0
seaborn>=0.13.0